Move character coordinates calculator to utils (from `InlineAutocomplete`)
